Update for super elevation, distant mountains (remove hard-coded areas) and multiplayer.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@1532 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/RunActivity/Simulator/Simulator.cs
@@ -87,7 +87,7 @@ namespace ORTS
         public bool BreakCouplers;
         public int CarVibrating;
         public int UseSuperElevation; //amount of superelevation
-        public float SuperElevationMinLen = 50f;
+        public int SuperElevationMinLen = 50;
         public float SuperElevationGauge = 1.435f;//1.435 guage
         // Used in save and restore form
         public string PathName = "<unknown>";
@@ -114,7 +114,7 @@ namespace ORTS
             BreakCouplers = Settings.BreakCouplers;
             CarVibrating = 0; //0 no vib, 1-2 mid vib, 3 max vib
             UseSuperElevation = Settings.UseSuperElevation;
-            SuperElevationMinLen = (float) Settings.SuperElevationMinLen;
+            SuperElevationMinLen = Settings.SuperElevationMinLen;
             SuperElevationGauge = (float)Settings.SuperElevationGauge/1000f;//gauge transfer from mm to m
             RoutePath = Path.GetDirectoryName(Path.GetDirectoryName(activityPath));
             RoutePathName = Path.GetFileName(RoutePath);
